Fix: Compatible version for Expo Go
Since there are only 3 specific versions that Expo Go is compatible with, I changed the app.json and package.json. Previously, I was modifying the runtimeVersion, but it was supposed to be the sdkVersion. I tested it on my android and it worked, but it was slow to establish the connection to download the app