feat: Markers are finally visible on the map
Markers are visible on the map, so that Is progress. I had to add the permissions again and use an async call. Then I had to get the markers asynchronously. I am going to pare down the UI a bit, but I will make a new branch in case of issues